requirements.txt
Django==4.2.7
graphene-django==3.1.5
django-filter==23.3
python-decouple==3.8
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.7.0


